+++ templates/zerver/api/changelog.md
@@ -11,6 +11,11 @@ below features are supported.
 
 ## Changes in Zulip 5.0
 
+**Feature level 76**
+
+* [`GET /realm/linkifiers`](/api/get-linkifiers): Added `render_format_string` field
+  to linkifier objects sent in `realm_linkifiers` events.
+
 **Feature level 75**
 
 * [`POST /register`](/api/register-queue), `PATCH /realm`: Replaced `allow_community_topic_editing`